pig/requirements.txt
Django>=1.11.7,<1.12
arrow>=0.10.0,<0.11
Celery>=4.1.0,<4.2
djangorestframework>=3.7.3,<3.8
django-auth-ldap>=1.2,<1.3
django-extensions>=1.9.7,<1.10
django-cors-headers>=2.1,<2.2
django-nose>=1.4.5,<1.5
django-reversion>=2.0.10,<2.1
psycopg2>=2.7.3,<2.8
redis>=2.10.6,<2.11
celery[redis]
